clean & document the routes
