+++ server/README.md
@@ -1,0 +1,20 @@
+# SlugFeed Server
+
+This is the backend server for Slugfeed. It's purpose is to handle api requests to the database and user authorization.
+
+## Directories
+- db: the database schema and seed files
+- routes: all visitable routes
+
+## Tech Stack
+- Node.js
+- Express
+- PostgreSQL
+- Knex
+
+## Installation, Seeding, and Startup
+```
+$ npm install
+$ npm run reDb
+$ npm start
+```
