+++ server/routes/README.md
@@ -1,0 +1,3 @@
+# SlugFeed Server/Routes
+
+Route handlers for api requests to the database and user authorization.
