+++ server/routes/_template.js
@@ -1,9 +1,18 @@
+/**
+ * Template for custom api routes
+ * Relies on dependency injection to get the router and knex database connection.
+ */
 "use strict";
 
 function appendRoutes(router, knex) {
 
-  router.get('/template', (request, response) => {
-    response.render('template');
+  router.get('/api/templates', (request, response) => {
+    knex
+      .select('*')
+      .from('templates')
+      .orderBy('created_at');
+      .then(templates => response.json(templates))
+      .catch(error => response.send(error));
   })
 
 };
