+++ server/routes/api-comments.js
@@ -1,20 +1,22 @@
+/** Custom routes for /api/comments */
 "use strict";
 const ensureLoggedIn = require('connect-ensure-login').ensureLoggedIn();
 const sentiment = require('sentiment');
 const sanitize = require('sanitizer').sanitize;
 
+/**
+ * Provide an initial score to the comment based on sentiment analysis.
+ * This is to encourage civility between commentors.
+ *
+ * The user is included in the function signature in case we want to
+ * make changes for habitually good/bad commenters.
+ */
 function initialScore(comment, user) {
-  /**
-   * Provide an initial score to the comment based on sentiment analysis.
-   * This is to encourage civility between commentors.
-   *
-   * The user is included in the function signature in case we want to
-   * make changes for habitually good/bad commenters.
-   */
   let score = Math.ceil(sentiment(comment).comparative * 3 / 5);
   return score;
 }
 
+/** Utility Query Def for getting the total of the votes on a given comment(s) */
 function baseQuery(knex, id) {
   let query = knex
     .select(knex.raw('comments.*, users.profile, SUM(CASE comment_votes.is_upvote WHEN true THEN 1 WHEN false THEN -1 ELSE 0 END) AS votes_score'))
@@ -22,9 +24,7 @@ function baseQuery(knex, id) {
     .leftJoin('comment_votes', 'comments.id', '=', 'comment_votes.comment_id')
     .leftJoin('users', 'comments.user_id', '=', 'users.id');
 
-  if (id) {
-    query.where('comments.id', id);
-  }
+  if (id) { query.where('comments.id', id); }
   query
     .groupBy('comments.id')
     .groupBy('users.id')
@@ -33,8 +33,10 @@ function baseQuery(knex, id) {
   return query;
 }
 
+/** The actual routing changes */
 function appendRoutes(router, knex) {
 
+  /** Custom GET table route incorporating the comment score */
   router.get(`/api/comments`, (request, response) => {
     const dbQuery = baseQuery(knex);
     for (const field of Object.keys(request.query)) {
@@ -49,15 +51,15 @@ function appendRoutes(router, knex) {
       .catch(error => response.status(422).send(error.detail));
   });
 
+  /** Custom GET record route incorporating the comment score */
   router.get(`/api/comments/:id`, (request, response) => {
     baseQuery(knex, request.params.id)
       .then(results => response.json(results))
       .catch(error => response.status(422).send(error.detail));
   });
 
-  router.post('/api/comments', (request, response) => {
-    console.log("HIT THE ROUTE");
-    // router.post('/api/comments', ensureLoggedIn, (request, response) => {
+  /** Custom POST record route calculating the initial score based on sentiment */
+  router.post('/api/comments', ensureLoggedIn, (request, response) => {
     const score = initialScore(request.body.content, request.user)
 
     return knex('comments')
